Fixes re. namespace trouble, hence rename well.py and cube.py
