+++ src/xtgeo/cube/__init__.py
@@ -3,4 +3,4 @@
 from __future__ import division, absolute_import
 from __future__ import print_function
 
-from xtgeo.cube.cube import Cube  # noqa
+from xtgeo.cube.cube1 import Cube  # noqa
