+++ src/xtgeo/well/__init__.py
@@ -4,7 +4,7 @@ from __future__ import division, absolute_import
 from __future__ import print_function
 
 # flake8: noqa
-from .well import Well
+from .well1 import Well
 from .wells import Wells
 
 from .blocked_well import BlockedWell
