+++ src/xtgeo/well/well1.py
